requirements.txt
Flask==2.0.2
numpy==1.19.2
keras==2.6.0
Pillow==8.4.0
tensorflow==2.6.0
waitress~=2.0.0